@@ -1,6 +1,6 @@
 {
   "name": "axis-discovery-ssdp",
-  "version": "5.0.0",
+  "version": "6.0.0",
   "description": "A Node.js SSDP (UPnP) client library written in TypeScript capable of searching for Axis Communication cameras.",
   "keywords": [
     "axis",
@@ -11,13 +11,13 @@
     "ssdp",
     "upnp"
   ],
-  "homepage": "https://github.com/FantasticFiasco/axis-discovery-ssdp",
+  "homepage": "https://github.com/FantasticFiasco/axis-discovery-ssdp-js",
   "repository": {
     "type": "git",
-    "url": "https://github.com/FantasticFiasco/axis-discovery-ssdp.git"
+    "url": "https://github.com/FantasticFiasco/axis-discovery-ssdp-js.git"
   },
   "bugs": {
-    "url": "https://github.com/FantasticFiasco/axis-discovery-ssdp/issues"
+    "url": "https://github.com/FantasticFiasco/axis-discovery-ssdp-js/issues"
   },
   "author": "Mattias Kindborg <mattias.kindborg@gmail.com> (https://twitter.com/FantasticFiasco)",
   "license": "Apache-2.0",
@@ -36,29 +36,31 @@
     "lint": "tslint \"./src/**/*.ts\" --exclude \"./**/*.d.ts\""
   },
   "dependencies": {
-    "@fantasticfiasco/expect": "1.0.0",
-    "debug": "3.1.0",
-    "request": "2.83.0",
-    "xml2js": "0.4.19"
+    "@fantasticfiasco/expect": "1.*",
+    "debug": "4.*",
+    "request": ">2.68.0",
+    "xml2js": "0.*"
   },
   "devDependencies": {
-    "yarn": "^1.3.2",
-    "@types/chai": "4.1.2",
-    "@types/debug": "0.0.30",
-    "@types/mocha": "2.2.48",
-    "@types/node": "9.4.6",
-    "@types/request": "2.47.0",
-    "@types/sinon": "4.1.3",
-    "@types/xml2js": "0.4.2",
-    "chai": "4.1.2",
-    "coveralls": "2.13.3",
-    "cpx": "1.5.0",
-    "istanbul": "0.4.5",
-    "mocha": "5.0.1",
-    "remap-istanbul": "0.10.1",
-    "rimraf": "2.6.2",
-    "sinon": "4.3.0",
-    "tslint": "^5.9.1",
-    "typescript": "2.6.2"
+    "@types/debug": "4.1.5",
+    "@types/jest": "26.0.16",
+    "@types/node": "14.14.10",
+    "@types/request": "2.48.5",
+    "@types/xml2js": "0.4.7",
+    "coveralls": "3.1.0",
+    "jest": "26.6.3",
+    "prettier": "2.2.1",
+    "rimraf": "3.0.2",
+    "ts-jest": "26.4.4",
+    "tslint": "6.1.3",
+    "typescript": "4.1.2"
+  },
+  "resolutions": {
+    "diff": ">=3.5.0",
+    "handlebars": ">=4.3.0",
+    "js-yaml": ">=3.13.0",
+    "lodash": ">=4.17.13",
+    "minimist": ">=1.2.2",
+    "sshpk": ">=1.13.2"
   }
 }
