@@ -1,6 +1,5 @@
 import * as expect from '@fantasticfiasco/expect';
-import * as request from 'request';
-
+import { get } from 'request';
 import { IHttpClient } from './IHttpClient';
 
 export class HttpClient implements IHttpClient {
@@ -8,7 +7,7 @@ export class HttpClient implements IHttpClient {
         expect.toExist(url);
 
         return new Promise<string>((resolve, reject) => {
-            request.get(url, undefined, (error: any, _, body: string) => {
+            get(url, undefined, (error: any, _, body: string) => {
                 if (error) {
                     reject(error);
                 } else {
