@@ -1,14 +1,11 @@
-import * as xml2js from 'xml2js';
+import { parseString } from 'xml2js';
 
 export class RootDescription {
-    private constructor(
-        readonly remoteAddress: string,
-        private readonly rootDescription: any) {
-    }
+    private constructor(readonly remoteAddress: string, private readonly rootDescription: any) {}
 
     public static parse(remoteAddress: string, xml: string): Promise<RootDescription> {
         return new Promise<RootDescription>((resolve, reject) => {
-            xml2js.parseString(xml, (error, result) => {
+            parseString(xml, (error, result) => {
                 if (error) {
                     reject(error);
                 } else {
