@@ -1,8 +1,7 @@
-import { setTimeout } from 'timers';
-import { Device, Discovery } from './';
+import { Device, Discovery } from './'
 
 // https://github.com/visionmedia/debug
-import * as debug from 'debug';
+import * as debug from 'debug'
 debug.enable('*');
 
 const discovery = new Discovery();
@@ -10,6 +9,7 @@ const discovery = new Discovery();
 discovery.onHello((device: Device) => {
         console.log(`${new Date().toLocaleTimeString()} - Hello from ${device.friendlyName} on ${device.address}`);
         console.log(`\tmodel: ${device.modelName}`);
+        console.log(`\tport: ${device.port}`);
         // console.log('Full data', JSON.stringify(device));
 });
 
@@ -17,10 +17,4 @@ discovery.onGoodbye((device: Device) => {
     console.log(`${new Date().toLocaleTimeString()} - Goodbye from ${device.macAddress} on ${device.address}`);
 });
 
-discovery.start()
-    // .then(() => discovery.search('urn:schemas-upnp-org:device:MediaServer:1'));
-    .then(() => discovery.search());
-
-setTimeout(() => {
-    discovery.stop();
-}, 3000);
+discovery.start().then(() => discovery.search());
