@@ -1,5 +1,5 @@
-import * as dgram from 'dgram';
-
+import * as expect from '@fantasticfiasco/expect';
+import { AddressInfo } from 'net';
 import { log } from '../logging';
 import { SSDP_MULTICAST_ADDRESS, SSDP_PORT } from './Constants';
 import { Message } from './Message';
@@ -26,29 +26,29 @@ export class MSearchSocket extends SocketBase {
         const message = new MSearch(target).toBuffer();
 
         return new Promise<void>((resolve, reject) => {
-            this.socket.send(
-                message,
-                0,
-                message.length,
-                SSDP_PORT,
-                SSDP_MULTICAST_ADDRESS,
-                (error: Error) => {
-                    if (error) {
-                        log('MSearchSocket#search - %o', error);
-                        reject(error);
-                    } else {
-                        resolve();
-                    }
-                },
-            );
+            if (!this.socket) {
+                reject(new Error('Socket has never been started'));
+                return;
+            }
+
+            this.socket.send(message, 0, message.length, SSDP_PORT, SSDP_MULTICAST_ADDRESS, (error: Error | null) => {
+                if (error) {
+                    log('MSearchSocket#search - %o', error);
+                    reject(error);
+                } else {
+                    resolve();
+                }
+            });
         });
     }
 
     protected onListening() {
-        log('MSearchSocket#onListening - %s:%d', this.socket.address().address, this.socket.address().port);
+        expect.toExist(this.socket, 'M-SEARCH socket has never been started');
+
+        log('MSearchSocket#onListening - %s:%d', (this.socket!.address() as AddressInfo).address, (this.socket!.address() as AddressInfo).port);
     }
 
-    protected onMessage(messageBuffer: Buffer, remote: dgram.AddressInfo) {
+    protected onMessage(messageBuffer: Buffer, remote: AddressInfo) {
         const message = new Message(remote.address, messageBuffer);
 
         if (message.method !== 'HTTP/1.1 200 OK') {
@@ -59,8 +59,10 @@ export class MSearchSocket extends SocketBase {
     }
 
     protected bind(): Promise<void> {
+        expect.toExist(this.socket, 'M-SEARCH socket has never been started');
+
         return new Promise<void>((resove) => {
-            this.socket.bind(undefined, this.address, () => resove());
+            this.socket!.bind(undefined, this.address, () => resove());
         });
     }
 }
