@@ -1,86 +1,73 @@
-import * as chai from 'chai';
 import * as os from 'os';
-import * as sinon from 'sinon';
-
+import { mocked } from 'ts-jest/utils';
 import { getIPv4Addresses } from './../../src/network-interfaces/NetworkInterface';
 import {
-    NETWORK_INTERFACE_WITH_TWO_ADDRESSES,
     NETWORK_INTERFACES_WITH_INTERNAL_ADDRESSES,
     NETWORK_INTERFACES_WITH_IPV6_ADDRESSES,
-    NETWORK_INTERFACES_WITH_TWO_ADDRESSES    ,
-    NO_NETWORK_INTERFACES
+    NETWORK_INTERFACES_WITH_TWO_ADDRESSES,
+    NETWORK_INTERFACE_WITH_TWO_ADDRESSES,
+    NO_NETWORK_INTERFACES,
 } from './NetworkInterface.mock';
 
-chai.should();
+jest.mock('os');
 
-describe('NetworkInterface', function() {
+describe('NetworkInterface', () => {
+    describe('#getIPv4Addresses', () => {
+        const osMock = mocked(os);
 
-    describe('#getIPv4Addresses', function() {
-
-        let osStub: sinon.SinonStub;
-
-        afterEach(function() {
-            osStub.restore();
-        });
-
-        it('should return addresses from one network interface', function() {
+        test('should return addresses from one network interface', () => {
             // Arrange
-            osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(NETWORK_INTERFACE_WITH_TWO_ADDRESSES);
+            osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACE_WITH_TWO_ADDRESSES);
 
             // Act
             const addresses = getIPv4Addresses();
 
             // Assert
-            addresses.should.be.eql(['1.1.1.1', '2.2.2.2']);
+            expect(addresses).toEqual(['1.1.1.1', '2.2.2.2']);
         });
 
-        it('should return addresses from multiple network interfaces', function() {
+        test('should return addresses from multiple network interfaces', () => {
             // Arrange
-            osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(NETWORK_INTERFACES_WITH_TWO_ADDRESSES);
+            osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACES_WITH_TWO_ADDRESSES);
 
             // Act
             const addresses = getIPv4Addresses();
 
             // Assert
-            addresses.should.be.eql(['1.1.1.1', '2.2.2.2']);
+            expect(addresses).toEqual(['1.1.1.1', '2.2.2.2']);
         });
 
-        it('should not return internal addresses', function() {
+        test('should not return internal addresses', () => {
             // Arrange
-            osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(NETWORK_INTERFACES_WITH_INTERNAL_ADDRESSES);
+            osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACES_WITH_INTERNAL_ADDRESSES);
 
             // Act
             const addresses = getIPv4Addresses();
 
             // Assert
-            addresses.should.be.empty;
+            expect(Object.keys(addresses)).toHaveLength(0);
         });
 
-        it('should not return IPv6 addresses', function() {
+        test('should not return IPv6 addresses', () => {
             // Arrange
-            osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(NETWORK_INTERFACES_WITH_IPV6_ADDRESSES);
+            osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACES_WITH_IPV6_ADDRESSES);
 
             // Act
             const addresses = getIPv4Addresses();
 
             // Assert
-            addresses.should.be.empty;
+            expect(Object.keys(addresses)).toHaveLength(0);
         });
 
-        it('should not fail on systems without network interfaces', function() {
+        test('should not fail on systems without network interfaces', () => {
             // Arrange
-            osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(NO_NETWORK_INTERFACES);
+            osMock.networkInterfaces.mockReturnValue(NO_NETWORK_INTERFACES);
 
             // Act
             const addresses = getIPv4Addresses();
 
             // Assert
-            addresses.should.be.empty;
+            expect(Object.keys(addresses)).toHaveLength(0);
         });
     });
 });
